Added support for linting es6 modules
